Disable conceal when editing Markdown
It's done by `polyglot`.  For example, ``` for code blocks gets
concealed (hidden except when cursor is on that line).
I don't like it.

The vimrc edit is supposed to disable conceal for polyglot's markdown
plugin, but it doesn't work.
So, I added a file in after/ to disable the whole conceal feature when
vim is in Markdown filetype.